[hotfix] refactoring for deploying
